requirements.txt
streamlit>=1.28.0
crewai>=0.28.0
python-dotenv>=1.0.0
pysqlite3-binary>=0.5.0